requirements.txt
python-fasthtml

connectorx
pyarrow
polars